Fixed errors with Rubocop
