+++ my_list.rb
@@ -1,0 +1,14 @@
+require_relative 'my_enumerable'
+class MyList
+  include my_enumerable
+
+  attr_reader :list
+
+  def initialize(*elements)
+    @list = elements
+  end
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
